# optimize io (70%+ startup speedups)