--- a/lua/nvim-lsp-installer/platform.lua
+++ b/lua/nvim-lsp-installer/platform.lua
@@ -11,6 +11,7 @@ M.arch = arch_aliases[uname.machine] or uname.machine
 M.is_win = vim.fn.has "win32" == 1
 M.is_unix = vim.fn.has "unix" == 1
 M.is_mac = vim.fn.has "mac" == 1
+M.is_linux = not M.is_mac and M.is_unix
 
 -- PATH separator
 M.path_sep = M.is_win and ";" or ":"