--- a/lua/nvim-lsp-installer/servers/bashls/init.lua
+++ b/lua/nvim-lsp-installer/servers/bashls/init.lua
@@ -1,13 +1,13 @@
 local server = require "nvim-lsp-installer.server"
 local npm = require "nvim-lsp-installer.installers.npm"
 
-local root_dir = server.get_server_root_path "bash"
-
-return server.Server:new {
-    name = "bashls",
-    root_dir = root_dir,
-    installer = npm.packages { "bash-language-server@latest" },
-    default_options = {
-        cmd = { npm.executable(root_dir, "bash-language-server"), "start" },
-    },
-}
+return function(name, root_dir)
+    return server.Server:new {
+        name = name,
+        root_dir = root_dir,
+        installer = npm.packages { "bash-language-server@latest" },
+        default_options = {
+            cmd = { npm.executable(root_dir, "bash-language-server"), "start" },
+        },
+    }
+end