--- a/lua/nvim-lsp-installer/servers/clangd/init.lua
+++ b/lua/nvim-lsp-installer/servers/clangd/init.lua
@@ -4,21 +4,23 @@ local platform = require "nvim-lsp-installer.platform"
 local Data = require "nvim-lsp-installer.data"
 local std = require "nvim-lsp-installer.installers.std"
 
-local root_dir = server.get_server_root_path "clangd"
-
 local VERSION = "12.0.1"
 
 local target = Data.coalesce(
     Data.when(platform.is_mac, "clangd-mac-%s.zip"),
-    Data.when(platform.is_unix, "clangd-linux-%s.zip"),
+    Data.when(platform.is_linux, "clangd-linux-%s.zip"),
     Data.when(platform.is_win, "clangd-windows-%s.zip")
 ):format(VERSION)
 
-return server.Server:new {
-    name = "clangd",
-    root_dir = root_dir,
-    installer = std.unzip_remote(("https://github.com/clangd/clangd/releases/download/%s/%s"):format(VERSION, target)),
-    default_options = {
-        cmd = { path.concat { root_dir, ("clangd_%s"):format(VERSION), "bin", "clangd" } },
-    },
-}
+return function(name, root_dir)
+    return server.Server:new {
+        name = name,
+        root_dir = root_dir,
+        installer = std.unzip_remote(
+            ("https://github.com/clangd/clangd/releases/download/%s/%s"):format(VERSION, target)
+        ),
+        default_options = {
+            cmd = { path.concat { root_dir, ("clangd_%s"):format(VERSION), "bin", "clangd" } },
+        },
+    }
+end