--- a/lua/nvim-lsp-installer/servers/cmake/init.lua
+++ b/lua/nvim-lsp-installer/servers/cmake/init.lua
@@ -1,13 +1,13 @@
 local server = require "nvim-lsp-installer.server"
 local pip3 = require "nvim-lsp-installer.installers.pip3"
 
-local root_dir = server.get_server_root_path "cmake"
-
-return server.Server:new {
-    name = "cmake",
-    root_dir = root_dir,
-    installer = pip3.packages { "cmake-language-server" },
-    default_options = {
-        cmd = { pip3.executable(root_dir, "cmake-language-server") },
-    },
-}
+return function(name, root_dir)
+    return server.Server:new {
+        name = name,
+        root_dir = root_dir,
+        installer = pip3.packages { "cmake-language-server" },
+        default_options = {
+            cmd = { pip3.executable(root_dir, "cmake-language-server") },
+        },
+    }
+end