--- a/lua/nvim-lsp-installer/servers/cssls/init.lua
+++ b/lua/nvim-lsp-installer/servers/cssls/init.lua
@@ -1,3 +1,1 @@
-local create_server = require "nvim-lsp-installer.servers.vscode-langservers-extracted"
-
-return create_server("cssls", "vscode-css-language-server")
+return require "nvim-lsp-installer.servers.vscode-langservers-extracted" "vscode-css-language-server"