--- a/lua/nvim-lsp-installer/servers/diagnosticls/init.lua
+++ b/lua/nvim-lsp-installer/servers/diagnosticls/init.lua
@@ -1,13 +1,13 @@
 local server = require "nvim-lsp-installer.server"
 local npm = require "nvim-lsp-installer.installers.npm"
 
-local root_dir = server.get_server_root_path "diagnosticls"
-
-return server.Server:new {
-	name = "diagnosticls",
-	root_dir = root_dir,
-	installer = npm.packages { "diagnostic-languageserver" },
-	default_options = {
-		cmd = { npm.executable(root_dir, "diagnostic-languageserver"), "--stdio" },
-	},
-}
+return function(name, root_dir)
+    return server.Server:new {
+        name = name,
+        root_dir = root_dir,
+        installer = npm.packages { "diagnostic-languageserver" },
+        default_options = {
+            cmd = { npm.executable(root_dir, "diagnostic-languageserver"), "--stdio" },
+        },
+    }
+end