--- a/lua/nvim-lsp-installer/servers/dockerls/init.lua
+++ b/lua/nvim-lsp-installer/servers/dockerls/init.lua
@@ -1,13 +1,13 @@
 local server = require "nvim-lsp-installer.server"
 local npm = require "nvim-lsp-installer.installers.npm"
 
-local root_dir = server.get_server_root_path "dockerfile"
-
-return server.Server:new {
-    name = "dockerls",
-    root_dir = root_dir,
-    installer = npm.packages { "dockerfile-language-server-nodejs@latest" },
-    default_options = {
-        cmd = { npm.executable(root_dir, "docker-langserver"), "--stdio" },
-    },
-}
+return function(name, root_dir)
+    return server.Server:new {
+        name = name,
+        root_dir = root_dir,
+        installer = npm.packages { "dockerfile-language-server-nodejs@latest" },
+        default_options = {
+            cmd = { npm.executable(root_dir, "docker-langserver"), "--stdio" },
+        },
+    }
+end