--- a/lua/nvim-lsp-installer/servers/efm/init.lua
+++ b/lua/nvim-lsp-installer/servers/efm/init.lua
@@ -1,13 +1,13 @@
 local server = require "nvim-lsp-installer.server"
 local go = require "nvim-lsp-installer.installers.go"
 
-local root_dir = server.get_server_root_path "efm"
-
-return server.Server:new {
-    name = "efm",
-    root_dir = root_dir,
-    installer = go.packages { "github.com/mattn/efm-langserver" },
-    default_options = {
-        cmd = { go.executable(root_dir, "efm-langserver") },
-    },
-}
+return function(name, root_dir)
+    return server.Server:new {
+        name = name,
+        root_dir = root_dir,
+        installer = go.packages { "github.com/mattn/efm-langserver" },
+        default_options = {
+            cmd = { go.executable(root_dir, "efm-langserver") },
+        },
+    }
+end