… README.md
… doc/nvim-lsp-installer.txt
… lua/nvim-lsp-installer/servers/elmls/init.lua
… lua/nvim-lsp-installer/servers/ember/init.lua
… lua/nvim-lsp-installer/servers/eslintls/init.lua
… lua/nvim-lsp-installer/servers/fortls/init.lua
… lua/nvim-lsp-installer/servers/gopls/init.lua
… lua/nvim-lsp-installer/servers/graphql/init.lua
… lua/nvim-lsp-installer/servers/groovyls/init.lua
… lua/nvim-lsp-installer/servers/hls/init.lua
… lua/nvim-lsp-installer/servers/html/init.lua
… lua/nvim-lsp-installer/servers/init.lua
… lua/nvim-lsp-installer/servers/intelephense/init.lua
… lua/nvim-lsp-installer/servers/jedi_language_server/init.lua
… lua/nvim-lsp-installer/servers/jsonls/init.lua
… lua/nvim-lsp-installer/servers/kotlin_language_server/init.lua
… lua/nvim-lsp-installer/servers/ocamlls/init.lua
… lua/nvim-lsp-installer/servers/omnisharp/init.lua
… lua/nvim-lsp-installer/servers/prismals/init.lua
… lua/nvim-lsp-installer/servers/purescriptls/init.lua
… lua/nvim-lsp-installer/servers/pylsp/init.lua
… lua/nvim-lsp-installer/servers/pyright/init.lua
… lua/nvim-lsp-installer/servers/rescriptls/init.lua
… lua/nvim-lsp-installer/servers/rome/init.lua
… lua/nvim-lsp-installer/servers/rust_analyzer/init.lua
… lua/nvim-lsp-installer/servers/solargraph/init.lua
… lua/nvim-lsp-installer/servers/sqlls/init.lua
… lua/nvim-lsp-installer/servers/sqls/init.lua
… lua/nvim-lsp-installer/servers/stylelint_lsp/init.lua
… lua/nvim-lsp-installer/servers/sumneko_lua/init.lua