So I've done some work to defer require calls as much as possible. These are by far the biggest culprits in terms of performance. On my Windows machine a single require took ~5ms (multiply that for each available server...). The problem today is that all the metadata for a given server exist exclusively inside each server's Lua module - which forces us to require all servers to do things like identify which ones are installed etc. I've started separating this a bit and colocating all the essential metadata in one location, and deferring all the "ad-hoc" metadata to the server modules themselves. This allows us to very quickly identify which servers are installed by scanning the root install directory once (via libuv), and then only require the installed servers.

The important metric is what's called when a user registers a on_server_ready callback, because that's the first touchpoint. Here are some profiling results I did across 3 different machines and operating systems:

```
macOS (11 servers installed)
37ms -> 10ms (~73% faster)

macOS (1 server installed)
37ms -> 1ms (~97% faster)

Windows (11 servers installed)
230ms -> 48ms (~80% faster)

Windows (0 servers installed)
230ms -> 0ms (∞% faster)

Low-end Ubuntu machine (11 servers installed)
80ms -> 24ms (~70% faster)
```